serverobject 아닌 파일 위치 이동
