+++ app/[locale]/academics/undergraduate/degree-requirements/page.tsx
@@ -1,10 +1,10 @@
-import { getMockDegreeRequirements } from '@/data/serverObjects';
+import { getDegreeRequirements } from '@/apis/academicsServer';
 
 import DegreeRequirementsBody from '@/components/academics/DegreeRequirements';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 
 export default async function UndergraduteDegreeRequirementsPage() {
-  const data = await getMockDegreeRequirements();
+  const data = await getDegreeRequirements();
 
   return (
     <PageLayout titleType="big" titleMargin="mb-9">
