+++ app/[locale]/research/groups/page.tsx
@@ -1,5 +1,3 @@
-export const dynamic = 'force-dynamic';
-
 import { getResearchGroups } from '@/apis/research';
 
 import SelectionList from '@/components/common/selection/SelectionList';
